Notify sysgit deploymentsd watcher to create a new deployments.json on paasta start/stop/restart
Previously this was handled by a gitolite hook, but that went away when services moved off gitolite. This is the same approach used by mark-for-deployment; we just missed that we needed it on start/stop/restart as well.

See PAASTA-16512
